+++ tests/markovchain.py
@@ -1,6 +1,8 @@
 import pickle
 import unittest
 
+import numpy as np
+
 from hopsy import *
 
 ProposalTypes = [
@@ -67,3 +69,31 @@ class MarkovChainTests(unittest.TestCase):
             mc = MarkovChain(problem, proposal)
             dump = pickle.dumps(mc)
             new_mc = pickle.loads(dump)
+
+            self.assertTrue(np.all(np.isclose(mc.state, new_mc.state)))
+
+    def test_markovchain_pickling_with_equality_constraints(self):
+        problem = Problem([[1, 1, 0], [-1, 0, 0], [0, -1, 0]], [1, 0, 0])
+        S = np.array([[2, 1, 0]])
+        h = np.array([0.5])
+
+        problem = add_box_constraints(problem, lower_bound=-10, upper_bound=10)
+        problem = add_equality_constraints(problem, A_eq=S, b_eq=h)
+        problem.starting_point = compute_chebyshev_center(problem)
+
+        for ProposalType in ProposalTypes:
+            try:
+                proposal = ProposalType(problem)
+            except:
+                pass
+            mc = MarkovChain(problem, proposal)
+            dump = pickle.dumps(mc)
+            new_mc = pickle.loads(dump)
+
+            self.assertTrue(np.all(np.isclose(mc.state, new_mc.state)))
+            # tests setting of state with equality constraints. This triggers a transform internally,
+            # which introduced a bug in the past. Therefore, we test it specifically.
+            mc.state = mc.state
+            mc.state = new_mc.state
+            new_mc.state = mc.state
+            self.assertTrue(np.all(np.isclose(mc.state, new_mc.state)))